app/requirements.txt
# =======================
# File: requirements.txt
# Path: requirements.txt
# =======================

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.3
google-generativeai==0.3.1
python-dotenv==1.0.0
Pillow==10.1.0